==> 03_LPV_Climate_Analysis.Rmd <==
---
title: "LPV_Culture_Analysis"
author: "Chad H. Hochberg"
date: "`r Sys.Date()`"
output: html_document
---

#Analytic File Relies on Having Run
#1a: Survey Analysis Files (Not Run in CrunchR)
#1b. K23_Aim1_LPV_Cohort_ETL.Rmd [Latest Version]
#2.  K23_Aim1_LPV_Cohort_MV_Volume.Rmd [Latest Version]
#3.  K23_Aim1_LPV_Cohort_CohortClean.Rmd [Latest Version]
```{r setup, include=FALSE}
library(tidyverse) #dplyr syntax for data wrangling
library(data.table) #Some commands here are faster than dplyr when working with big data
library(arrow) #This require allows working with apache arrow which is a memory efficient way to work with these data
library(collapse)#Fast Data Manipulation Package
library(DescTools) #Help for Descriptive Statistics
library(tableone) #Makes Useful Descriptive Tables
library(ggpubr) #Arrange plots
library(viridis) #Colorblind Friendly Color Maps
library(glmmTMB) #Package for Running Multi-level Models
library(marginaleffects)


#Set Working Directory
knitr::opts_knit$set(root.dir = "~/workspace/Storage/chochbe1/persistent/K23_Aim1_LPV_Analysis")

#Create Commonly Used Function
`%!in%` = Negate(`%in%`)

#Project Path
project_path <- '~/workspace/Storage/chochbe1/persistent/K23_Aim1_LPV_Analysis'
```


```{r Open Analytic Data, include=FALSE}
lpv_analytic <- open_dataset(paste0(project_path, '/data/lpv_analytic.parquet')) |>
  collect()
```


```{r When Did These Admissions Happen}
cat('There are', length(unique(lpv_analytic$cohort_id)), 'patients in this analysis \n')

#What Dates Are Represented Here?
lpv_analytic <- lpv_analytic |> arrange(final_admit_date) 
#First by Hospital Admission Dates
cat('\n First Hospital Admission Date:')
first(lpv_analytic$final_admit_date)
cat('\n Last Hospital Admission Date:')
last(lpv_analytic$final_admit_date)

#Now by Study ICU Admission
#Here Use 'unit_in_dttm' which is the time in to the target ICU
lpv_analytic <- lpv_analytic |> arrange(unit_in_dttm) 
cat('\n First ICU Admission Date:')
first(lpv_analytic$unit_in_dttm)
cat('\n Last ICU Admission Date:')
last(lpv_analytic$unit_in_dttm)

#How Many Patients are Still in the Hospital at the Time of Database Lock?
#NOTE: Confirmed that Missing Discharge Date in This Cohort Means Still Hospitalized at Time of Last Data Pull
cat('\n How Many Patients Were Still in the Hospital at the Time of Database Lock?')
print(sum(is.na(lpv_analytic$final_dc_date)))

#Histogram of Admission By Study Month

#What ICUs Are Represented Here?
icu_num <- lpv_analytic |>
  group_by(study_icu) |>
  summarise(number_admits=n(),
            percent = n()/dim(lpv_analytic)[1]) |>
  mutate(percent=round(percent, digits = 3)) |>
  arrange(-number_admits)
icu_num

#Bar Chart of Admissions Per Study ICU
ggplot(icu_num, aes(x=reorder(study_icu, number_admits), y=number_admits)) +
  geom_col(width = 0.95) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1)) +
  scale_y_continuous(breaks=seq(0,500, by=50), limits = c(0,500)) +
  xlab('ICUs') + 
  ylab('Number of Admissions')
ggsave('admits_by_unit.pdf',
       path = 'graphs/')
  
#Admissions Per Month
lpv_analytic <- lpv_analytic |>
  mutate(admit_month=month(final_admit_date)) |>
  mutate(admit_year=year(final_admit_date)) |>
  mutate(admit_month=fcase(
    admit_year==2022, admit_month-6,
    admit_year==2023, admit_month+6,
    admit_year==2024, admit_month+18
  ))
ggplot(lpv_analytic, aes(x=admit_month, fill=hospital)) +
  geom_bar(position = 'stack') +
  theme_classic() +
  scale_x_continuous(breaks=seq(1,24, by=1), limits = c(0, 25),
                     labels = c('Jul22', 'Aug22', 'Sep22', 'Oct22', 
                              'Nov22', 'Dec22', 'Jan23', 'Feb23', 'Mar23', 
                              'Apr23', 'May23', 'Jun23', 'Jul23', 'Aug23',
                              'Sep23', 'Oct23', 'Nov23', 'Dec23', 'Jan24',
                              'Feb24', 'Mar24', 'Apr24', 'May24', 'Jun24'),
                     name = 'Month-Year') +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1)) +
  scale_y_continuous(breaks=seq(0,150, by=25),
                     name = 'Number of Admissions') +
  scale_fill_viridis_d()
ggsave('admits_by_month.pdf',
       path = 'graphs/')
```
```{r Add a Few Additional Variables}
lpv_analytic <- lpv_analytic |>
  mutate(female=fifelse(gender=='Female', 1, 0)) |>
  relocate(female, .after = gender) |>
  mutate(admit_category=fcase(
    ed_to_icu==1, 'ED',
    elective_surg==1 | ed_to_or==1, 'Elective Surg',
    direct_icu==1, 'Outside Hospital',
    default = 'Ward')) |>
  mutate(icu_type=fcase(
    study_icu %in% c('BMC MEDICAL ICU',
                     'JHH ZAYED 10E',
                     'JHH WEINBERG 5C'),
    'Medical',
    study_icu %in% c('BMC CARDIAC ICU', 
                    'JHH ZAYED 5W', 
                    'JHH ZAYED 3W'), 
    'Specialty',
    study_icu %in% c('BMC SURGICAL ICU',
                     'JHH WEINBERG 3A',
                     'JHH ZAYED 9E'), 
    'Surgical',
    study_icu %in% c('JHH ZAYED 5E'),
    'CV Surgery',
    study_icu %in% c('HCGH 3C ICU', 
                     'SH 3100 INTENSIVE CARE',
                     'SH 3400 INTENSIVE CARE',
                     'SMH ICU'), 'Mixed'
  )) |>
  mutate(icu_beds=fcase(
    study_icu=='BMC CARDIAC ICU', 12,
    study_icu=='BMC MEDICAL ICU', 12,
    study_icu=='BMC SURGICAL ICU', 12,
    study_icu=='HCGH 3C ICU', 16,
    study_icu=='JHH WEINBERG 3A', 24,
    study_icu=='JHH WEINBERG 5C', 10, 
    study_icu=='JHH ZAYED 10E', 24,
    study_icu=='JHH ZAYED 3W', 20,
    study_icu=='JHH ZAYED 5E', 18,
    study_icu=='JHH ZAYED 5W', 12,
    study_icu=='JHH ZAYED 9E', 12,
    study_icu=='SH 3100 INTENSIVE CARE', 12,
    study_icu=='SH 3400 INTENSIVE CARE', 12,
    study_icu=='SMH ICU', 16
  ))
```

```{r Who Are These Patients - Table 1}
to_tab <- c('age_at_admission', 
         'gender',
         'female',
         'race', 
         'ethnicity', 
         'nonwhite',
         'admit_year',
         'hospital',
         'icu_type',
         'icu_beds',
         'mv_volume_month',
         'transfer_to_jhhs', 
         'ed_to_icu',
         'ed_to_or',
         'elective_surg',
         'direct_icu',
         'admit_category',
         'elixhauser_count',
         'code_stat_time0',
         'CMR_Index_Mortality',
         'ccsr_system_cat',
         'dx_category',
         'covid_ccsr_dx',
         'covid_icd10_poa',
         'height_inches',
         'weight_kg',
         'bmi',
         'max_nr_sofa',
         'pf_ratio',
         'sf_ratio',
         'pf_fio2',
         'sf_fio2',
         'resp_compliance_baseline',
         'first_vent_mode',
         'received_ecmo',
         'inhosp_death',
         'dc_hospice')

factors_tab <- c(
         'gender',
         'female',
         'race', 
         'ethnicity', 
         'nonwhite',
         'admit_year',
         'hospital',
         'icu_type',
         'transfer_to_jhhs', 
         'admit_category',
         'ed_to_icu',
         'ed_to_or',
         'elective_surg',
         'direct_icu',
         'ccsr_system_cat',
         'dx_category',
         'covid_ccsr_dx',
         'covid_icd10_poa',
         'first_vent_mode',
         'received_ecmo',
         'code_stat_time0',
         'inhosp_death',
         'dc_hospice')

nonnorm_tab <-c('age_at_admission', 
          'icu_beds',
         'mv_volume_month',
         'elixhauser_count',
         'CMR_Index_Mortality',
         'height_inches',
         'weight_kg',
         'bmi',
         'max_nr_sofa',
         'pf_ratio',
         'sf_ratio',
         'pf_fio2',
         'sf_fio2',
         'resp_compliance_baseline')

#First Create Table with Numbers of Missing and Min/Max
tab1 <- CreateTableOne(vars = to_tab, data=lpv_analytic, factorVars = factors_tab)
summary(tab1)
#Now Create a Traditional Table 1, Using Non-parametric Testing
tab1 <- CreateTableOne(vars = to_tab, strata="study_icu", data=lpv_analytic, factorVars = factors_tab, addOverall = TRUE)
print(tab1, nonnormal=nonnorm_tab)
tab1_excel <- print(tab1, nonnorm=nonnorm_tab, printToggle = FALSE)
write.csv(tab1_excel, file="tables/table1_by-icu.csv")
rm(tab1_excel)
```

```{r Describe and Visualize Ventilator Measures - Table2ish}
#Create a 'Vent Hours' Metric for Whole Cohort Description 
lpv_analytic <- lpv_analytic |>
  mutate(lpv_hours_day1=sum(time_lpv_eligible24, na.rm = TRUE)) |>
  mutate(lpv_hours_day2=sum(time_lpv_eligible48, na.rm = TRUE)) |>
  mutate(lpv_hours_day3=sum(time_lpv_eligible72, na.rm = TRUE)) |>
  mutate(tv_hours_all=sum(time_tv_eligible24, na.rm = TRUE)) |>
  #Given Distribution (a lot of 0 or 1 in terms of % LPV) Create Several Binary Variables
  mutate(zero_percent_lpv=fcase(
    !is.na(percent_lpv) & percent_lpv==0, 1,
    !is.na(percent_lpv) & percent_lpv>0, 0,
    default = NA)) |>
  #Count Number oF Patients Included In Each Day
  mutate(n_day1=sum(!is.na(percent_lpv))) |>
  mutate(n_day2=sum(!is.na(percent_lpv_day2))) |>
  mutate(n_day3=sum(!is.na(percent_lpv_day3))) |>
  relocate(lpv_hours_day1:n_day3, .before = first_vent_mode) 

#Table the Ventilator Measures Overall and By ICU
to_tab <- names(lpv_analytic)[which(names(lpv_analytic) == "n_vent_observations"):which(names(lpv_analytic) == "peep_weighted_day3")]
factors_tab <- c('first_vent_mode', 'most_mode_tv_eligible', 'zero_percent_lpv')
nonnorm_tab <- names(lpv_analytic)[which(names(lpv_analytic)=='n_vent_observations'):which(names(lpv_analytic)=='tv_hours_all')]
#Add Day 2
nonnorm_tab <- c(nonnorm_tab, names(lpv_analytic)[which(names(lpv_analytic)=='n_vent_observations_day2'):which(names(lpv_analytic)=='peep_weighted_day3')])

#First Create Table with Numbers of Missing and Min/Max
tab2 <- CreateTableOne(vars = to_tab, data=lpv_analytic, factorVars = factors_tab)
summary(tab2)
#Now Create a Traditional Table 1, Using Non-parametric Testing
tab2 <- CreateTableOne(vars = to_tab, strata="study_icu", data=lpv_analytic, factorVars = factors_tab, addOverall = TRUE)
print(tab2, nonnormal=nonnorm_tab)
tab2_excel <- print(tab2, nonnorm=nonnorm_tab, printToggle = FALSE)
write.csv(tab2_excel, file="tables/table2_vent-outcomes_by-icu.csv")
rm(tab2_excel)

#Remake This Table but With Means and Standard Deviations
tab2_excel <- print(tab2, printToggle = FALSE)
write.csv(tab2_excel, file="tables/table2_vent-outcomes_by-icu_means.csv")
rm(tab2_excel, tab2)

#Visualize Distributions
#Prepare Dataset
df <- lpv_analytic |>
  mutate(tv_weighted24=fifelse(tv_weighted24>12, 12, tv_weighted24)) |>
  group_by(study_icu) |>
  mutate(temp_mean_tv=mean(tv_weighted24, na.rm = TRUE)) |>
  ungroup()

#Histogram of Hours at LPV
ggplot(df, aes(x=hours_at_lpv)) +
  geom_histogram(binwidth = 1) +
  scale_x_continuous(breaks=seq(0,24, by=2), limits = c(-1,25), name = 'Hours of Adherence to LPV in 1st 24 Hours') +
  theme_minimal()

#Percent LPV Adherence in 1st 24 Hours
ggplot(df, aes(x=percent_lpv)) +
  geom_histogram(binwidth = 0.1) +
  scale_x_continuous(breaks=seq(0,1, by=0.1), limits = c(-0.1,1.1), name = '% Adherence to LPV in 1st 24 Hours') +
  theme_minimal()

#TV Weighted by Unit
ggplot(df, aes(x=reorder(study_icu, temp_mean_tv), y=tv_weighted24)) +
  stat_summary(fun = mean, color = "orangered4",
               geom = "crossbar", width = 0.60, alpha = 0.75, linewidth = 0.50) +
  geom_dotplot(binaxis = "y", stackdir = "center", dotsize = 0.60,
               binwidth = 0.05, alpha = 0.80, 
               position=position_jitter(0.05)) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1)) +
  scale_y_continuous(breaks=seq(4,12, by=1), 
                     labels=c('4', '5', '6', '7', '8', '9', '10', '11', '12+'),
                     name = 'Tidal Volume (kg/IBW) \n Weighted Average in 1st 24 Hours') +
  xlab('ICU') +
  scale_fill_viridis_d()
ggsave('dotplot_tv24_weighted.pdf',
       path = 'graphs/')

#Cumulative TV
#First Plot Weighted TV
#Use Empirical Cumulative Distribution Function
#Find the Percentiles at X=6.5 and x=8
# Create the ECDF function
ecdf_function <- ecdf(df$tv_weighted24)
# Get the ECDF value at x = 6.5
ecdf_6.5 <- ecdf_function(6.5)
# Get the ECDF value at x = 8.0
ecdf_8.0 <- ecdf_function(8.0)

ggplot(df, aes(x=tv_weighted24)) +
  stat_ecdf(geom = 'line') +
  theme_minimal() +
  scale_x_continuous(breaks=seq(4, 12, by=1), limits=c(3.5, 12),
                     labels=c('4', '5', '6', '7', '8', '9', '10', '11', '12+'), 
                     name='Tidal Volume Per Predicted Body Weight') +
  ylab('Cumulative Frequency') +
  #Shaded Region at <=8
  annotate("rect", xmin = 3.5, xmax = 8.0, ymin = 0, ymax = ecdf_8.0,
           alpha = .1,fill = "red") +
  annotate("text", x=6.0, y=0.78, label = 'TV<8.0 ml/kg') +
  #Shaded Region at <=6.5
  annotate("rect", xmin = 3.5, xmax = 6.5, ymin = 0, ymax = ecdf_6.5,
           alpha = .5,fill = "lightblue") +
  annotate("text", x=4.5, y=0.4, label = 'TV<6.5 ml/kg') 
ggsave('cumulative_freq_tvweighted.pdf',
       path = 'graphs/')

#Now Grouped by Study ICU
ggplot(df, aes(x=tv_weighted24, color=study_icu)) +
  stat_ecdf(geom = 'line') +
  theme_minimal() +
  scale_color_viridis_d('ICU') +
   scale_x_continuous(breaks=seq(4, 12, by=1), limits=c(3.5, 12),
                     labels=c('4', '5', '6', '7', '8', '9', '10', '11', '12+'), 
                     name='Tidal Volume Per Predicted Body Weight') +
  ylab('Cumulative Frequency') 
ggsave('cumulative_freq_tvweighted_byicu.pdf',
       path = 'graphs/')

#Now Evaluate Pplat
df <- df |>
  group_by(study_icu) |>
  mutate(temp_mean_pplat=mean(pplat_weighted24, na.rm = TRUE)) |>
  ungroup()
ggplot(df, aes(x=reorder(study_icu, temp_mean_pplat), y=pplat_weighted24)) +
  stat_summary(fun = mean, color = "orangered4",
               geom = "crossbar", width = 0.60, alpha = 0.75, linewidth = 0.50) +
  geom_dotplot(binaxis = "y", stackdir = "center", dotsize = 0.60,
               binwidth = 0.25, alpha = 0.80, 
               position=position_jitter(0.05)) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1)) +
  ylab('Plateau Pressure (cmH2O)') +
  xlab('ICU')
ggsave('dotplot_pplat24_byicu.pdf',
       path = 'graphs/')

#Driving Pressure
df <- df |>
  group_by(study_icu) |>
  mutate(temp_mean_dp=mean(dp_weighted24, na.rm = TRUE)) |>
  ungroup()
ggplot(df, aes(x=reorder(study_icu, temp_mean_dp), y=dp_weighted24)) +
  stat_summary(fun = mean, color = "orangered4",
               geom = "crossbar", width = 0.60, alpha = 0.75, linewidth = 0.50) +
  geom_dotplot(binaxis = "y", stackdir = "center", dotsize = 0.60,
               binwidth = 0.25, alpha = 0.80, 
               position=position_jitter(0.05)) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1)) +
  ylab('Driving Pressure (cmH2O)') +
  xlab('ICU')
ggsave('dotplot_dp24_byicu.pdf',
       path = 'graphs/')

#Finally look at Percent Time at LPV and LPV Restricted
#First LPV
df <- lpv_analytic |>
  mutate(tv_weighted24=fifelse(tv_weighted24>12, 12, tv_weighted24)) |>
  group_by(study_icu) |>
  mutate(temp_mean_tv=mean(tv_weighted24, na.rm = TRUE)) |>
  mutate(mean_lpv=mean(percent_lpv, na.rm=TRUE)*10) |> #Put on Same Scale as TV for Graph
  ungroup() 
ggplot(df, aes(x=reorder(study_icu, temp_mean_tv), y=tv_weighted24)) +
  stat_summary(fun = mean, color = "orangered4",
               geom = "crossbar", width = 0.60, alpha = 0.75, linewidth = 0.50) +
  geom_dotplot(binaxis = "y", stackdir = "center", dotsize = 0.60, color = 'lightgrey',
               binwidth = 0.05, alpha = 0.80, 
               position=position_jitter(0.05)) +
  theme_minimal() +
  geom_point(aes(x=study_icu, y=mean_lpv), color = 'darkblue') +
  geom_line(aes(y = mean_lpv), color = "darkblue", size = 1, group = 1) +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1)) +
  scale_y_continuous(breaks=seq(1,12, by=1), 
                     labels=c('1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12+'),
                     name = 'Tidal Volume (kg/IBW) \n Weighted Average in 1st 24 Hours',
      sec.axis =  sec_axis(trans = ~.*10, 
      name = '     % of First 24 Hours LPV Acheived',
      breaks=seq(0,100, by = 10))) +
  xlab('ICU')
ggsave('percent_LPV_byicu.pdf',
       path = 'graphs/')

#Version for Paper
#Version for Paper
ggplot(df, aes(x=reorder(study_icu, temp_mean_tv), y=tv_weighted24)) +
  stat_summary(fun = mean, color = "orangered4",
               geom = "crossbar", width = 0.60, alpha = 0.75, linewidth = 0.50) +
  geom_dotplot(binaxis = "y", stackdir = "center", dotsize = 0.60, color = 'lightgrey',
               binwidth = 0.05, alpha = 0.80, 
               position=position_jitter(0.05)) +
  theme_minimal() +
  geom_point(aes(x=study_icu, y=mean_lpv), color = 'darkblue') +
  geom_line(aes(y = mean_lpv), color = "darkblue", size = 0.75, group = 1) +
  scale_y_continuous(breaks=seq(1,12, by=1), 
                     labels=c('1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12+'),
                     name = 'Tidal Volume (kg/IBW) \n Weighted Average in 1st 24 Hours',
                     sec.axis =  sec_axis(trans = ~.*10, 
                                          name = '        Percent LPV Acheived (%)',
                                          breaks=seq(0,100, by = 10))) +
  scale_x_discrete(labels=c('1','2', '3','4','5','6','7','8',
                            '9','10','11','12','13','14')) +
  xlab('ICUs Ranked by Mean Time-weighted Tidal Volume') +
  theme(
    axis.text.y = element_text(color = "orangered4"),  # Color for primary y-axis ticks
    axis.text.y.right = element_text(color = "darkblue"),  # Color for secondary y-axis ticks
    axis.title.y.right = element_text(color = "darkblue", size = 12), # Secondary y-axis title color
    axis.title.y = element_text(color = "orangered4", size = 12),  # Primary y-axis title color
    panel.grid.major = element_line(color = "grey80", size = 0.15), # Lighter major gridlines
    panel.grid.minor = element_line(color = "grey90", size = 0.075),  # Lighter minor gridlines
  )
ggsave('percent_LPV_byicu_forpaper.pdf',
       path = 'graphs/')

#Now LPV Restricted
df <- lpv_analytic |>
  mutate(tv_weighted24=fifelse(tv_weighted24>12, 12, tv_weighted24)) |>
  group_by(study_icu) |>
  mutate(temp_mean_tv=mean(tv_weighted24, na.rm = TRUE)) |>
  mutate(mean_lpv=mean(percent_lpv_restrict, na.rm=TRUE)*10) |> #Put on Same Scale as TV for Graph
  ungroup() 

#Make Chart
ggplot(df, aes(x=reorder(study_icu, temp_mean_tv), y=tv_weighted24)) +
  stat_summary(fun = mean, color = "orangered4",
               geom = "crossbar", width = 0.60, alpha = 0.75, linewidth = 0.50) +
  geom_dotplot(binaxis = "y", stackdir = "center", dotsize = 0.60, color = 'lightgrey',
               binwidth = 0.05, alpha = 0.80, 
               position=position_jitter(0.05)) +
  theme_minimal() +
  geom_point(aes(x=study_icu, y=mean_lpv), color = 'darkblue') +
  geom_line(aes(y = mean_lpv), color = "darkblue", size = 1, group = 1) +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1)) +
  scale_y_continuous(breaks=seq(1,12, by=1), 
                     labels=c('1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12+'),
                     name = 'Tidal Volume (kg/IBW) \n Weighted Average in 1st 24 Hours',
      sec.axis =  sec_axis(trans = ~.*10, 
      name = '% of First 24 Hours LPV Acheived \n Restricted Definition',
      breaks=seq(0,100, by = 10))) +
  xlab('ICU')
ggsave('percent_LPV_restrict_byicu.pdf',
       path = 'graphs/')
rm(df, ecdf_6.5, ecdf_8.0)
```

```{r Did LPV adherence and Weighted TV Change Over Time}
#NOTE: This is a pre-specified check and if there is a change over time primary analysis, will only include the last 6 months of 2024

#Plot LPV Percent OVer Time
df <- lpv_analytic |>
  group_by(study_month) |>
  mutate(mean_lpv=mean(percent_lpv, na.rm = TRUE)) |>
  mutate(mean_lpv_restrict=mean(percent_lpv_restrict, na.rm = TRUE)) |>
  mutate()
ggplot(df, aes(x=study_month, y=mean_lpv)) +
  geom_point() +
  geom_line() +
  geom_smooth(method  = 'lm') +
  ylab('% LPV Delivery, First 24 Hours') +
  scale_x_continuous(breaks=seq(1,25, by=1), limits = c(0, 25),
                     labels = c('Jul22', 'Aug22', 'Sep22', 'Oct22', 
                              'Nov22', 'Dec22', 'Jan23', 'Feb23', 'Mar23', 
                              'Apr23', 'May23', 'Jun23', 'Jul23', 'Aug23',
                              'Sep23', 'Oct23', 'Nov23', 'Dec23', 'Jan24',
                              'Feb24', 'Mar24', 'Apr24', 'May24', 'Jun24', 'Aug24'),
                     name = 'Month-Year') +
  scale_y_continuous(breaks=seq(0,1, by=0.1), labels = scales::percent, limits = c(0,1.0)) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1)) +
  theme(panel.grid.minor.x = element_blank()) +
  theme(panel.grid.minor.y = element_blank())
ggsave('lpv_by_month.pdf',
       path = 'graphs/')
  
ggplot(df, aes(x=study_month, y=mean_lpv_restrict)) +
  geom_point() +
  geom_line() +
  geom_smooth(method  = 'lm') +
  ylab('% LPV Delivery, First 24 Hours \n [Restrictive Definition]') +
  scale_x_continuous(breaks=seq(1,25, by=1), limits = c(0, 25),
                     labels = c('Jul22', 'Aug22', 'Sep22', 'Oct22', 
                              'Nov22', 'Dec22', 'Jan23', 'Feb23', 'Mar23', 
                              'Apr23', 'May23', 'Jun23', 'Jul23', 'Aug23',
                              'Sep23', 'Oct23', 'Nov23', 'Dec23', 'Jan24',
                              'Feb24', 'Mar24', 'Apr24', 'May24', 'Jun24', 'Aug24'),
                     name = 'Month-Year') +
  scale_y_continuous(breaks=seq(0,1, by=0.1), labels = scales::percent, limits = c(0,1.0)) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1)) +
  theme(panel.grid.minor.x = element_blank()) +
  theme(panel.grid.minor.y = element_blank())

#Now Run Mixed Model with ICU as Random Effect, LPV Hours as Outcome with total time as offest, Study Month as Predictor)
#Form is Negative Binomial
df <- lpv_analytic |> 
  mutate(hours_at_lpv=round(hours_at_lpv, digits = 0)) |>
  mutate(hours_at_lpv_restrict=round(hours_at_lpv_restrict, digits = 0)) |>
  filter(time_lpv_eligible24>0) 
month_mod <- glmmTMB(hours_at_lpv ~ study_month + offset(log(time_lpv_eligible24)) +
                       (1 | study_icu),
                     data = df, 
                     family=nbinom2)
summary(month_mod)
check_overdispersion(month_mod)

#Repeat with Restricted
month_mod <- glmmTMB(hours_at_lpv_restrict ~ study_month + offset( log(time_lpv_eligible24)) + (1 | study_icu),
                     data = df, 
                     family=nbinom2)
summary(month_mod)
check_overdispersion(month_mod)

#Multi-level Logistic Model of Zero Percent LPV by Study Month
zero_lpv_month <- glmmTMB(zero_percent_lpv ~ study_month  + (1 | study_icu),
                     data = df, 
                     family=binomial)
summary(zero_lpv_month)

#Time Trends are Not Significant, Will Move to Use Full Sample in Analysis
```

```{r Visualize Association of LPV and Survey Measures}
#Graph LPV Percent and Survey Score
#Prepare Dataset 
df <- lpv_analytic |> 
  group_by(study_icu) |>
  mutate(mean_lpv=mean(percent_lpv, na.rm = TRUE)) |>
  mutate(mean_lpv_restrict=mean(percent_lpv_restrict, na.rm = TRUE)) |>
  ungroup()

#Association of Mean Unit Culture and LPV
culture <- ggplot(df) +
  geom_smooth(aes(x=mean_unit_culture, y=percent_lpv), method = 'lm', color ='black') +
  geom_jitter(aes(x=mean_unit_culture, y=percent_lpv, color = reorder(study_icu, -mean_unit_culture)), 
              width = 0.001, height = 0.05, size = 0.9, alpha = 0.8) +
  scale_y_continuous(labels = scales::percent) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ICU Implementation Culture Score') +
  ylab('% of Time LPV Delivered, First 24 Hours') +
  guides(colour = guide_legend(override.aes = list(size=3), title = 'ICU'))
print(culture)

#Association of Mean Unit Stress and LPV
stress <- ggplot(df) +
  geom_smooth(aes(x=mean_unit_stress, y=percent_lpv), method = 'lm', color ='black') +
  geom_jitter(aes(x=mean_unit_stress, y=percent_lpv, color = reorder(study_icu, -mean_unit_culture)), 
              width = 0.001, height = 0.05, size = 0.9, alpha = 0.8) +
  scale_y_continuous(labels = scales::percent) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ICU Staff Stress Score') +
  ylab('% of Time LPV Delivered, First 24 Hours') +
  guides(colour = guide_legend(override.aes = list(size=3), title = 'ICU'))
print(stress)

#Association of Mean Unit Resources and LPV
resources <- ggplot(df) +
  geom_smooth(aes(x=mean_unit_resources, y=percent_lpv), method = 'lm', color ='black') +
  geom_jitter(aes(x=mean_unit_resources, y=percent_lpv, color = reorder(study_icu, -mean_unit_culture)), 
              width = 0.001, height = 0.05, size = 0.9, alpha = 0.8) +
  scale_y_continuous(labels = scales::percent) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ICU Resources Score') +
  ylab('% of Time LPV Delivered, First 24 Hours') +
  guides(colour = guide_legend(override.aes = list(size=3), title = 'ICU'))
print(resources)

#Association of Mean Unit ORIC Score
oric <- ggplot(df) +
  geom_smooth(aes(x=mean_unit_oric, y=percent_lpv), method = 'lm', color ='black') +
  geom_jitter(aes(x=mean_unit_oric, y=percent_lpv, color = reorder(study_icu, -mean_unit_culture)), 
              width = 0.001, height = 0.05, size = 0.9, alpha = 0.8) +
  scale_y_continuous(labels = scales::percent) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ORIC Score') +
  ylab('% of Time LPV Delivered, First 24 Hours') +
  guides(colour = guide_legend(override.aes = list(size=3), title = 'ICU'))
print(oric)

#Association of Mean Unit Efficacy Score
efficacy <- ggplot(df) +
  geom_smooth(aes(x=mean_unit_efficacy, y=percent_lpv), method = 'lm', color ='black') +
  geom_jitter(aes(x=mean_unit_efficacy, y=percent_lpv, color = reorder(study_icu, -mean_unit_culture)), 
              width = 0.001, height = 0.05, size = 0.9, alpha = 0.8) +
  scale_y_continuous(labels = scales::percent) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('Mean Unit Self-Efficacy Score') +
  ylab('% of Time LPV Delivered, First 24 Hours') +
  guides(colour = guide_legend(override.aes = list(size=3), title = 'ICU'))
print(efficacy)

#Arrange in 1 Graph Using ggpubr
#Modify and Arrange
culture <- culture + xlab('Culture Score') + ylab('')
stress <- stress + xlab('Stress Score') + ylab('')
resources <- resources + xlab('Resource Score') + ylab('% of Time LPV Delivered, First 24 Hours')
oric <- oric + xlab('ORIC Score') + ylab('')
efficacy <- efficacy + xlab('Self-Efficacy Score') + ylab('')
ggarrange(culture, stress, resources, oric, efficacy, ncol=2, nrow=3, common.legend = TRUE, legend="right")
ggsave('surveys_lpv_byicu.pdf',
       path = 'graphs/',
       width = 11,
       height = 8.5,
       units = 'in')

#Correlation Matrix
df <- df |> filter(!is.na(percent_lpv)) |> 
  select(percent_lpv, mean_unit_culture, mean_unit_stress, mean_unit_resources, mean_unit_oric, mean_unit_efficacy)
cat('\n Correlation Matrix for LPV (Restricted) and Survey Measures \n')
round(cor(df), digits = 2)
```

```{r Visualize Association of LPV Restricted and Survey Measures and Measure Correlation Coefficients}
df <- lpv_analytic |> 
  group_by(study_icu) |>
  mutate(mean_lpv=mean(percent_lpv, na.rm = TRUE)) |>
  mutate(mean_lpv_restrict=mean(percent_lpv_restrict, na.rm = TRUE)) |>
  mutate(icu_count=n()) |>
  ungroup() |>
  group_by(icu_type) |> 
  mutate(count_type=sum(icu_count)) |> 
  ungroup()

#Association of Mean Unit Culture and LPV
culture <- ggplot(df) +
  geom_smooth(aes(x=mean_unit_culture, y=percent_lpv_restrict), method = 'lm', color ='black') +
  geom_jitter(aes(x=mean_unit_culture, y=percent_lpv_restrict, color = reorder(study_icu, -mean_unit_culture)), 
              width = 0.001, height = 0.05, size = 0.9, alpha = 0.8) +
  scale_y_continuous(labels = scales::percent) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ICU Implementation Culture Score') +
  ylab('% of Time LPV Delivered, First 24 Hours \n (Restricted Definition)') +
  guides(colour = guide_legend(override.aes = list(size=3), title = 'ICU'))
print(culture)

#Association of Mean Unit Stress and LPV
stress <- ggplot(df) +
  geom_smooth(aes(x=mean_unit_stress, y=percent_lpv_restrict), method = 'lm', color ='black') +
  geom_jitter(aes(x=mean_unit_stress, y=percent_lpv_restrict, color = reorder(study_icu, -mean_unit_culture)), 
              width = 0.001, height = 0.05, size = 0.9, alpha = 0.8) +
  scale_y_continuous(labels = scales::percent) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ICU Staff Stress Score') +
  ylab('% of Time LPV Delivered, First 24 Hours \n (Restricted Definition)') +
  guides(colour = guide_legend(override.aes = list(size=3), title = 'ICU'))
print(stress)

#Association of Mean Unit Resources and LPV
resources <- ggplot(df) +
  geom_smooth(aes(x=mean_unit_resources, y=percent_lpv_restrict), method = 'lm', color ='black') +
  geom_jitter(aes(x=mean_unit_resources, y=percent_lpv_restrict, color = reorder(study_icu, -mean_unit_culture)), 
              width = 0.001, height = 0.05, size = 0.9, alpha = 0.8) +
  scale_y_continuous(labels = scales::percent) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ICU Resources Score') +
  ylab('% of Time LPV Delivered, First 24 Hours \n (Restricted Definition)') +
  guides(colour = guide_legend(override.aes = list(size=3), title = 'ICU'))
print(resources)

#Association of Mean Unit ORIC Score
oric <- ggplot(df) +
  geom_smooth(aes(x=mean_unit_oric, y=percent_lpv_restrict), method = 'lm', color ='black') +
  geom_jitter(aes(x=mean_unit_oric, y=percent_lpv_restrict, color = reorder(study_icu, -mean_unit_culture)), 
              width = 0.001, height = 0.05, size = 0.9, alpha = 0.8) +
  scale_y_continuous(labels = scales::percent) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ORIC Score') +
  ylab('% of Time LPV Delivered, First 24 Hours \n (Restricted Definition)') +
  guides(colour = guide_legend(override.aes = list(size=3), title = 'ICU'))
print(oric)

#Association of Mean Unit Efficacy Score
efficacy <- ggplot(df) +
  geom_smooth(aes(x=mean_unit_efficacy, y=percent_lpv_restrict), method = 'lm', color ='black') +
  geom_jitter(aes(x=mean_unit_efficacy, y=percent_lpv_restrict, color = reorder(study_icu, -mean_unit_culture)), 
              width = 0.001, height = 0.05, size = 0.9, alpha = 0.8) +
  scale_y_continuous(labels = scales::percent) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('Mean Unit Self-Efficacy Score') +
  ylab('% of Time LPV Delivered, First 24 Hours \n (Restricted Definition)') +
  guides(colour = guide_legend(override.aes = list(size=3), title = 'ICU'))
print(efficacy)

#Arrange in 1 Graph Using ggpubr
#Modify and Arrange
culture <- culture + xlab('Culture Score') + ylab('')
stress <- stress + xlab('Stress Score') + ylab('')
resources <- resources + xlab('Resource Score') + ylab('% of Time LPV Delivered, First 24 Hours \n (Restricted Definition)')
oric <- oric + xlab('ORIC Score') + ylab('')
efficacy <- efficacy + xlab('Self-Efficacy Score') + ylab('')
ggarrange(culture, stress, resources, oric, efficacy, ncol=2, nrow=3, common.legend = TRUE, legend="right")
ggsave('surveys_lpv-restrict_byicu.pdf',
       path = 'graphs/',
       width = 11,
       height = 8.5,
       units = 'in')

#Unit Level Plot of Mean Vent Outcome and Survey Score
#Culture Plot
ggplot(df) +
  geom_smooth(aes(x=mean_unit_culture, 
                  y=percent_lpv), 
              method = 'glm', alpha = 0.5) +
  geom_point(aes(x=mean_unit_culture, y=mean_lpv, 
                          size = icu_count, 
                 color = reorder(icu_type, -count_type)), 
             alpha = 0.8) +
  scale_y_continuous(breaks=seq(0,1, by=0.1), labels = scales::percent, limits = c(0,1.0)) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ICU Implementation Culture Score') +
  ylab('% of Time LPV Delivered, First 24 Hours') +
  guides(colour = guide_legend(override.aes = list(size=5), title = 'ICU Type'), size = 'none')
ggsave('scatter_bubble_lpv-culture.pdf',
       path = 'graphs/')

#Stress
ggplot(df) +
  geom_smooth(aes(x=mean_unit_stress, 
                  y=percent_lpv), 
              method = 'glm', alpha = 0.5) +
  geom_point(aes(x=mean_unit_stress, y=mean_lpv, 
                          size = icu_count, 
                 color = reorder(icu_type, -count_type)), 
             alpha = 0.8) +
  scale_y_continuous(breaks=seq(0,1, by=0.1), labels = scales::percent, limits = c(0,1.0)) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ICU Staff Stress Score') +
  ylab('% of Time LPV Delivered, First 24 Hours') +
  guides(colour = guide_legend(override.aes = list(size=5), title = 'ICU Type'), size = 'none')
ggsave('scatter_bubble_lpv-stress.pdf',
       path = 'graphs/')

#Resources
ggplot(df) +
  geom_smooth(aes(x=mean_unit_resources, 
                  y=percent_lpv), 
              method = 'glm', alpha = 0.5) +
  geom_point(aes(x=mean_unit_resources, y=mean_lpv, 
                          size = icu_count, 
                 color = reorder(icu_type, -count_type)), 
             alpha = 0.8) +
  scale_y_continuous(breaks=seq(0,1, by=0.1), labels = scales::percent, limits = c(0,1.0)) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('Mean Unit Resource Score') +
  ylab('% of Time LPV Delivered, First 24 Hours') +
  guides(colour = guide_legend(override.aes = list(size=5), title = 'ICU Type'), size = 'none')
ggsave('scatter_bubble_lpv-resources.pdf',
       path = 'graphs/')

#ORIC
ggplot(df) +
  geom_smooth(aes(x=mean_unit_oric, 
                  y=percent_lpv), 
              method = 'glm', alpha = 0.5) +
  geom_point(aes(x=mean_unit_oric, y=mean_lpv, 
                          size = icu_count, 
                 color = reorder(icu_type, -count_type)), 
             alpha = 0.8) +
  scale_y_continuous(breaks=seq(0,1, by=0.1), labels = scales::percent, limits = c(0,1.0)) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ORIC Score') +
  ylab('% of Time LPV Delivered, First 24 Hours') +
  guides(colour = guide_legend(override.aes = list(size=5), title = 'ICU Type'), size = 'none')
ggsave('scatter_bubble_lpv-oric.pdf',
       path = 'graphs/')

#Efficacy
ggplot(df) +
  geom_smooth(aes(x=mean_unit_efficacy, 
                  y=percent_lpv), 
              method = 'glm', alpha = 0.5) +
  geom_point(aes(x=mean_unit_efficacy, y=mean_lpv, 
                          size = icu_count, 
                 color = reorder(icu_type, -count_type)), 
             alpha = 0.8) +
  scale_y_continuous(breaks=seq(0,1, by=0.1), labels = scales::percent, limits = c(0,1.0)) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('Mean Unit Efficacy Score') +
  ylab('% of Time LPV Delivered, First 24 Hours') +
  guides(colour = guide_legend(override.aes = list(size=5), title = 'ICU Type'), size = 'none')
ggsave('scatter_bubble_lpv-efficacy.pdf',
       path = 'graphs/')

#Make One Graph that Prints the Legend for the Bubble Size
ggplot(df) +
  geom_smooth(aes(x=mean_unit_culture, 
                  y=percent_lpv), 
              method = 'glm', alpha = 0.5) +
  geom_point(aes(x=mean_unit_culture, y=mean_lpv, 
                 size = icu_count, 
                 color = reorder(study_icu, -mean_unit_culture)), 
             alpha = 0.8) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ICU Implementation Culture Score') +
  ylab('% of Time LPV Delivered, First 24 Hours') +
  guides(colour = 'none', 
         size = guide_legend(title = 'Number of Patients'))
ggsave('scatter_bubble_bubble-size.pdf',
       path = 'graphs/')
  
#Correlation Matrix
df <- df |> filter(!is.na(percent_lpv_restrict)) |> 
  select(percent_lpv_restrict, mean_unit_culture, mean_unit_stress, mean_unit_resources, mean_unit_oric, mean_unit_efficacy)
cat('\n Correlation Matrix for LPV (Restricted) and Survey Measures \n')
round(cor(df), digits = 2)


```

```{r Survey and Tidal Volumes}
df <- lpv_analytic |> 
  group_by(study_icu) |>
  mutate(mean_tv=mean(tv_weighted24, na.rm = TRUE)) |>
  mutate(tv_weighted24=fifelse(tv_weighted24>12, 12, tv_weighted24)) |>
  mutate(icu_count=n()) |>
  ungroup()

#Association of Mean Unit Culture and LPV
culture <- ggplot(df) +
  geom_smooth(aes(x=mean_unit_culture, y=tv_weighted24), method = 'lm', color ='black') +
  geom_jitter(aes(x=mean_unit_culture, y=tv_weighted24, color = reorder(study_icu, -mean_unit_culture)), 
              width = 0.001, height = 0.05, size = 0.9, alpha = 0.8) +
  scale_y_continuous(breaks=seq(4,12, by=1), 
                     labels=c('4', '5', '6', '7', '8', '9', '10', '11', '12+')) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ICU Implementation Culture Score') +
  ylab('Tidal Volume (ml/kg-PBW)') +
  guides(colour = guide_legend(override.aes = list(size=3), title = 'ICU'))
print(culture)

#Association of Mean Unit Stress and LPV
stress <- ggplot(df) +
  geom_smooth(aes(x=mean_unit_stress, y=tv_weighted24), method = 'lm', color ='black') +
  geom_jitter(aes(x=mean_unit_stress, y=tv_weighted24, color = reorder(study_icu, -mean_unit_culture)), 
              width = 0.001, height = 0.05, size = 0.9, alpha = 0.8) +
  scale_y_continuous(breaks=seq(4,12, by=1), 
                     labels=c('4', '5', '6', '7', '8', '9', '10', '11', '12+')) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ICU Staff Stress Score') +
  ylab('Tidal Volume (ml/kg-PBW)') +
  guides(colour = guide_legend(override.aes = list(size=3), title = 'ICU'))
print(stress)

#Association of Mean Unit Resources and LPV
resources <- ggplot(df) +
  geom_smooth(aes(x=mean_unit_resources, y=tv_weighted24), method = 'lm', color ='black') +
  geom_jitter(aes(x=mean_unit_resources, y=tv_weighted24, color = reorder(study_icu, -mean_unit_culture)), 
              width = 0.001, height = 0.05, size = 0.9, alpha = 0.8) +
  scale_y_continuous(breaks=seq(4,12, by=1), 
                     labels=c('4', '5', '6', '7', '8', '9', '10', '11', '12+')) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ICU Resources Score') +
  ylab('Tidal Volume (ml/kg-PBW)') +
  guides(colour = guide_legend(override.aes = list(size=3), title = 'ICU'))
print(resources)

#Association of Mean Unit ORIC Score
oric <- ggplot(df) +
  geom_smooth(aes(x=mean_unit_oric, y=tv_weighted24), method = 'lm', color ='black') +
  geom_jitter(aes(x=mean_unit_oric, y=tv_weighted24, color = reorder(study_icu, -mean_unit_culture)), 
              width = 0.001, height = 0.05, size = 0.9, alpha = 0.8) +
  scale_y_continuous(breaks=seq(4,12, by=1), 
                     labels=c('4', '5', '6', '7', '8', '9', '10', '11', '12+')) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ORIC Score') +
  ylab('Tidal Volume (ml/kg-PBW)') +
  guides(colour = guide_legend(override.aes = list(size=3), title = 'ICU'))
print(oric)

#Association of Mean Unit Efficacy Score
efficacy <- ggplot(df) +
  geom_smooth(aes(x=mean_unit_efficacy, y=tv_weighted24), method = 'lm', color ='black') +
  geom_jitter(aes(x=mean_unit_efficacy, y=tv_weighted24, color = reorder(study_icu, -mean_unit_culture)), 
              width = 0.001, height = 0.05, size = 0.9, alpha = 0.8) +
  scale_y_continuous(breaks=seq(4,12, by=1), 
                     labels=c('4', '5', '6', '7', '8', '9', '10', '11', '12+')) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('Mean Unit Efficacy Score') +
  ylab('Tidal Volume (ml/kg-PBW)') +
  guides(colour = guide_legend(override.aes = list(size=3), title = 'ICU'))
print(culture)

#Arrange in 1 Graph Using ggpubr
#Modify and Arrange
culture <- culture + xlab('Culture Score') + ylab('Tidal Volume (ml/kg-PBW)')
stress <- stress + xlab('Stress Score') + ylab('')
resources <- resources + xlab('Resource Score') + ylab('Tidal Volume (ml/kg-PBW)')
oric <- oric + xlab('ORIC Score') + ylab('')
efficacy <- efficacy + xlab('Self-Efficacy Score') + ylab('Tidal Volume (ml/kg-PBW)')
ggarrange(culture, stress, resources, oric, efficacy, ncol=2, nrow=3, common.legend = TRUE, legend="right")
ggsave('surveys_tv_weighted24_byicu.pdf',
       path = 'graphs/',
       width = 11,
       height = 8.5,
       units = 'in')


#Make Bubble Chart For Each Survey Score and TV
#Unit Level Plot of Mean Vent Outcome and Survey Score
#Culture Plot
ggplot(df) +
  geom_smooth(aes(x=mean_unit_culture, 
                  y=tv_weighted24), 
              method = 'glm', alpha = 0.5) +
  geom_point(aes(x=mean_unit_culture, y=mean_tv, 
                 size = icu_count, 
                 color = reorder(study_icu, -mean_unit_culture)), 
             alpha = 0.8) +
  scale_y_continuous(breaks=seq(6,7.8, by=0.2), limits = c(6,7.8)) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ICU Implementation Culture Score') +
  ylab('Tidal Volume (ml/kg-PBW)') +
  guides(colour = guide_legend(override.aes = list(size=3), title = 'ICU'), size = 'none')
ggsave('scatter_bubble_tv-culture.pdf',
       path = 'graphs/')

#Stress
ggplot(df) +
  geom_smooth(aes(x=mean_unit_stress, 
                  y=tv_weighted24), 
              method = 'glm', alpha = 0.5) +
  geom_point(aes(x=mean_unit_stress, y=mean_tv, 
                 size = icu_count, 
                 color = reorder(study_icu, -mean_unit_culture)), 
             alpha = 0.8) +
  scale_y_continuous(breaks=seq(6,7.8, by=0.2), limits = c(6,7.8)) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ICU Staff Stress Score') +
  ylab('Tidal Volume (ml/kg-PBW)') +
  guides(colour = guide_legend(override.aes = list(size=3), title = 'ICU'), size = 'none')
ggsave('scatter_bubble_tv-stress.pdf',
       path = 'graphs/')

#Resources
ggplot(df) +
  geom_smooth(aes(x=mean_unit_resources, 
                  y=tv_weighted24), 
              method = 'glm', alpha = 0.5) +
  geom_point(aes(x=mean_unit_resources, y=mean_tv, 
                 size = icu_count, 
                 color = reorder(study_icu, -mean_unit_culture)), 
             alpha = 0.8) +
  scale_y_continuous(breaks=seq(6,7.8, by=0.2), limits = c(6,7.8)) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('Mean Unit Resources') +
  ylab('Tidal Volume (ml/kg-PBW)') +
  guides(colour = guide_legend(override.aes = list(size=3), title = 'ICU'), size = 'none')
ggsave('scatter_bubble_tv-resources.pdf',
       path = 'graphs/')

#ORIC
ggplot(df) +
  geom_smooth(aes(x=mean_unit_oric, 
                  y=tv_weighted24), 
              method = 'glm', alpha = 0.5) +
  geom_point(aes(x=mean_unit_oric, y=mean_tv, 
                 size = icu_count, 
                 color = reorder(study_icu, -mean_unit_culture)), 
             alpha = 0.8) +
  scale_y_continuous(breaks=seq(6,7.8, by=0.2), limits = c(6,7.8)) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ORIC Score') +
  ylab('Tidal Volume (ml/kg-PBW)') +
  guides(colour = guide_legend(override.aes = list(size=3), title = 'ICU'), size = 'none')
ggsave('scatter_bubble_tv-oric.pdf',
       path = 'graphs/')

#Efficacy
ggplot(df) +
  geom_smooth(aes(x=mean_unit_efficacy, 
                  y=tv_weighted24), 
              method = 'glm', alpha = 0.5) +
  geom_point(aes(x=mean_unit_efficacy, y=mean_tv, 
                          size = icu_count, 
                 color = reorder(study_icu, -mean_unit_culture)), 
             alpha = 0.8) +
  scale_y_continuous(breaks=seq(6,7.8, by=0.2), limits = c(6,7.8)) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('Mean Unit Efficacy Score') +
  ylab('% of Time LPV Delivered, First 24 Hours') +
  guides(colour = guide_legend(override.aes = list(size=3), title = 'ICU'), size = 'none')
ggsave('scatter_bubble_tv-efficacy.pdf',
       path = 'graphs/')
  
#Correlation Matrix
df <- df |> filter(!is.na(tv_weighted24)) |> 
  select(tv_weighted24, mean_unit_culture, mean_unit_stress, mean_unit_resources, mean_unit_oric, mean_unit_efficacy)
cat('\n Correlation Matrix for LPV (Restricted) and Survey Measures \n')
round(cor(df), digits = 2)

```

```{r Prepare For Outcome Modelling - Relationship of Covariates to Outcome}
#Covariates in Multivariable Model:
#Age, non-white, elixhauser mortality index, height, weight, COVID, nonresp SOFA, Vasoactive Infuion, S/F ratio, respiratory compliance, Admit type, primary diagnosis category, study month, hospital (as fixed effect?), MV Volume

df <- lpv_analytic

#Function to Create Histogram for Continous Covariates
fn.histo_associate <- function(df, var, outcome) {
  ggplot(df, 
       aes(x = var, y = outcome)) +
  geom_jitter(height = 0.05, width = 0.1, alpha = 0.25, size = 0.5) +
  geom_smooth() +
  labs(x = name, y = outcome_name, size = 0.1) +
  scale_y_continuous(labels = scales::percent) +
  theme(axis.title.x = element_text(size = 7),
    axis.title.y = element_text(size = 7)
  )
}
outcome_name <- "LPV(%)"
name <- "Age"
p1 <- fn.histo_associate(df, df$age_at_admission, df$percent_lpv)
name <- "Height"
p2 <- fn.histo_associate(df, df$height_inches, df$percent_lpv)
name <- "Weight"
p3 <- fn.histo_associate(df, df$weight_kg, df$percent_lpv)
name <- "Elixhauser Mortality Index"
p4 <- fn.histo_associate(df, df$CMR_Index_Mortality, df$percent_lpv)
name <- "S/F Ratio"
p5 <- fn.histo_associate(df, df$sf_ratio, df$percent_lpv)
name <- "P/F Ratio"
p6 <- fn.histo_associate(df, df$pf_ratio, df$percent_lpv)
name <- "Respiratory Compliance"
p7 <- fn.histo_associate(df,                      df$resp_compliance_baseline, df$percent_lpv)
name <- "ICU Monthly Volume"
p8 <- fn.histo_associate(df, df$mv_volume_month, df$percent_lpv)

ggarrange(p1, p2, p3, p4, p5, p6, p7, p8, ncol=2, nrow=4)
ggsave('covariate_lpv_percent.pdf',
       path = 'graphs/')

```
```{r Simple Ecologic Approach - Regress Mean Unit LPV Adherence on Mean Unit Survey Score}
#Scale Survey Scores by IQR (so interpretation of regression is for each 1 IQR increase)
lpv_analytic <- lpv_analytic |>
  mutate(mean_unit_culture_scaled=mean_unit_culture/sd(mean_unit_culture)) |>
  mutate(mean_unit_stress_scaled=mean_unit_stress/sd(mean_unit_stress)) |>
  mutate(mean_unit_resources_scaled=mean_unit_resources/sd(mean_unit_resources)) |>
  mutate(mean_unit_oric_scaled=mean_unit_oric/sd(mean_unit_oric)) |>
  mutate(mean_unit_efficacy_scaled=mean_unit_efficacy/sd(mean_unit_efficacy))

#Create Unit Level Data Frame
#Create Weights of % of Total Population
df <- lpv_analytic |> 
  mutate(total_n=n()) |>
  group_by(study_icu) |>
  mutate(mean_lpv=mean(percent_lpv, na.rm = TRUE)) |>
  mutate(mean_lpv_restrict=mean(percent_lpv_restrict, na.rm = TRUE)) |>
  mutate(icu_count=n()) |>
  mutate(weights=icu_count/total_n) |>
  filter(row_number()==1) |>
  ungroup()

#Weighted Linear Regression Using Simple Ecologic Approach to Get Test of Trend
culture <- lm(mean_lpv ~ mean_unit_culture_scaled, data = df, weights = weights)
summary(culture)
confint(culture)

stress <- lm(mean_lpv ~ mean_unit_stress_scaled, data = df, weights = weights)
summary(stress)
confint(stress)

resources <- lm(mean_lpv ~ mean_unit_resources_scaled, data = df, weights = weights)
summary(resources)
confint(resources)

efficacy <- lm(mean_lpv ~ mean_unit_efficacy_scaled, data = df, weights = weights)
summary(efficacy)
confint(efficacy)
```

```{r Univariable Mixed Effect Models with ICU as Random Effect}
#Create Function to Loop Through Individual Surveys
#Model form is of a rate using hours in negative binomial with time eligible as offset
fn.unvar_mle <- function(predictor, outcome, df) {
  glmmTMB(outcome ~ predictor + offset(log(time_lpv_eligible24)) +
                       (1 | study_icu),
                     data = df, 
                     family=nbinom2)
}
df <- lpv_analytic |>
  mutate(hours_at_lpv=round(hours_at_lpv, digits = 0))

#Now Run for Percent LPV
cat('ICU Culture Score \n \n')
mod1 <- fn.unvar_mle(df$mean_unit_culture_scaled, df$hours_at_lpv, df)
summary(mod1)
exp(confint(mod1))

cat('\n ICU Stress Score \n \n')
mod1 <- fn.unvar_mle(df$mean_unit_stress_scaled, df$hours_at_lpv, df)
summary(mod1)
exp(confint(mod1))

cat('\n ICU Resources \n \n')
mod1 <- fn.unvar_mle(df$mean_unit_resources_scaled, df$hours_at_lpv, df)
summary(mod1)
exp(confint(mod1))

cat('\n ICU ORIC \n \n')
mod1 <- fn.unvar_mle(df$mean_unit_oric_scaled, df$hours_at_lpv, df)
summary(mod1)
exp(confint(mod1))

cat('\n ICU Efficacy \n \n')
mod1 <- fn.unvar_mle(df$mean_unit_efficacy_scaled, df$hours_at_lpv, df)
summary(mod1)
exp(confint(mod1))
```


```{r Multivariable Models}
#Round Hours
df <- lpv_analytic |>
  mutate(hours_at_lpv=round(hours_at_lpv, digits = 0))

#Culture ICU
#Create the Multivariable Model Form
model_form <- hours_at_lpv ~ mean_unit_culture_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

culture_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(culture_mv_mod)
exp(confint(culture_mv_mod))

#ICU Stress
model_form <- hours_at_lpv ~ mean_unit_stress_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

stress_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(stress_mv_mod)
exp(confint(stress_mv_mod))

#ICU Resources
model_form <- hours_at_lpv ~ mean_unit_resources_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

resource_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(resource_mv_mod)
exp(confint(resource_mv_mod))

#ICU ORIC
model_form <- hours_at_lpv ~ mean_unit_oric_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

oric_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(oric_mv_mod)
exp(confint(oric_mv_mod))

#ICU Efficacy
model_form <- hours_at_lpv ~ mean_unit_efficacy_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

efficacy_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(efficacy_mv_mod)
exp(confint(efficacy_mv_mod))

```
```{r Repeat Multivariable Models for Restricted Time at LPV}
#Round Hours
df <- lpv_analytic |>
  mutate(hours_at_lpv_restrict=round(hours_at_lpv_restrict, digits = 0))

#Culture ICU
#Create the Multivariable Model Form
model_form <- hours_at_lpv_restrict ~ mean_unit_culture_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

culture_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(culture_mv_mod)
exp(confint(culture_mv_mod))

#ICU Stress
model_form <- hours_at_lpv_restrict ~ mean_unit_stress_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

stress_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(stress_mv_mod)
exp(confint(stress_mv_mod))

#ICU Resources
model_form <- hours_at_lpv_restrict ~ mean_unit_resources_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

resource_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(resource_mv_mod)
exp(confint(resource_mv_mod))

#ICU ORIC
model_form <- hours_at_lpv_restrict ~ mean_unit_oric_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

oric_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(oric_mv_mod)
exp(confint(oric_mv_mod))

#ICU Efficacy
model_form <- hours_at_lpv_restrict ~ mean_unit_efficacy_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

efficacy_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(efficacy_mv_mod)
exp(confint(efficacy_mv_mod))

```


```{r Secdonary Outcomes - Time Above TV of 8}
df <- lpv_analytic |>
  mutate(hours_tv_gt8=round(hours_tv_gt8), digits = 0)

#Culture ICU
#Create the Multivariable Model Form
model_form <- hours_tv_gt8 ~ mean_unit_culture_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_tv_eligible24)) + (1 | study_icu)

culture_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(culture_mv_mod)
exp(confint(culture_mv_mod))

#ICU Stress
model_form <- hours_tv_gt8 ~ mean_unit_stress_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_tv_eligible24)) + (1 | study_icu)

stress_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(stress_mv_mod)
exp(confint(stress_mv_mod))

#ICU Resources
model_form <- hours_tv_gt8 ~ mean_unit_resources_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_tv_eligible24)) + (1 | study_icu)

resource_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(resource_mv_mod)
exp(confint(resource_mv_mod))

#ICU ORIC
model_form <- hours_tv_gt8 ~ mean_unit_oric_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_tv_eligible24)) + (1 | study_icu)

oric_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(oric_mv_mod)
exp(confint(oric_mv_mod))

#ICU Efficacy
model_form <- hours_tv_gt8 ~ mean_unit_efficacy_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_tv_eligible24)) + (1 | study_icu)

efficacy_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(efficacy_mv_mod)
exp(confint(efficacy_mv_mod))

```


```{r Secondary Outcomes; Hour 24-48 (Day 2) LPV Adherence - Stratified Models}
#Round Hours
df <- lpv_analytic |>
  mutate(hours_at_lpv_day2=round(hours_at_lpv_day2, digits = 0))

#Culture ICU
#Create the Multivariable Model Form
model_form <- hours_at_lpv_day2 ~ mean_unit_culture_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible48)) + (1 | study_icu)

culture_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(culture_mv_mod)
exp(confint(culture_mv_mod))

#ICU Stress
model_form <- hours_at_lpv_day2 ~ mean_unit_stress_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible48)) + (1 | study_icu)

stress_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(stress_mv_mod)
exp(confint(stress_mv_mod))

#ICU Resources
model_form <- hours_at_lpv_day2 ~ mean_unit_resources_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible48)) + (1 | study_icu)

resource_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(resource_mv_mod)
exp(confint(resource_mv_mod))

#ICU ORIC
model_form <- hours_at_lpv_day2 ~ mean_unit_oric_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible48)) + (1 | study_icu)

oric_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(oric_mv_mod)
exp(confint(oric_mv_mod))

#ICU Efficacy
model_form <- hours_at_lpv_day2 ~ mean_unit_efficacy_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible48)) + (1 | study_icu)

efficacy_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(efficacy_mv_mod)
exp(confint(efficacy_mv_mod))
```
```{r Secondary Outcomes - Culture Measures and LPV Adherence on Day 3 (hours 48-72)}
#Round Hours
df <- lpv_analytic |>
  mutate(hours_at_lpv_day3=round(hours_at_lpv_day3, digits = 0))

#Culture ICU
#Create the Multivariable Model Form
model_form <- hours_at_lpv_day3 ~ mean_unit_culture_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible72)) + (1 | study_icu)

culture_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(culture_mv_mod)
exp(confint(culture_mv_mod))

#ICU Stress
model_form <- hours_at_lpv_day3 ~ mean_unit_stress_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible72)) + (1 | study_icu)

stress_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(stress_mv_mod)
exp(confint(stress_mv_mod))

#ICU Resources
model_form <- hours_at_lpv_day3 ~ mean_unit_resources_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible72)) + (1 | study_icu)

resource_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(resource_mv_mod)
exp(confint(resource_mv_mod))

#ICU ORIC
model_form <- hours_at_lpv_day3 ~ mean_unit_oric_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible72)) + (1 | study_icu)

oric_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(oric_mv_mod)
exp(confint(oric_mv_mod))

#ICU Efficacy
model_form <- hours_at_lpv_day3 ~ mean_unit_efficacy_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible72)) + (1 | study_icu)

efficacy_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(efficacy_mv_mod)
exp(confint(efficacy_mv_mod))
```

```{r Secondary Outcomes; LPV Adherence on Days 1-3 - Combined Mixed Effects Models with Interaction Term}
#Prepare One Dataset for Each Day 1-3
temp <- lpv_analytic |>
  mutate(hours_at_lpv_day1=round(hours_at_lpv, digits=0)) |>
  mutate(hours_at_lpv_day2=round(hours_at_lpv_day2, digits = 0)) |>
  mutate(hours_at_lpv_day3=round(hours_at_lpv_day3, digits = 0)) |>
  #Calculate Day 1, Day 2, Day3 Hours
  mutate(time_lpv_eligible_day1=time_lpv_eligible24) |>
  mutate(time_lpv_eligible_day2=time_lpv_eligible48) |>
  mutate(time_lpv_eligible_day3=time_lpv_eligible72)

day1 <- temp |>
  mutate(hours_at_lpv=hours_at_lpv_day1) |>
  mutate(time_lpv_eligible=time_lpv_eligible24) |>
  mutate(day=1)
day2 <- temp |>
  mutate(hours_at_lpv=hours_at_lpv_day2) |>
  mutate(time_lpv_eligible=time_lpv_eligible48) |>
  mutate(day=2)
day3 <- temp |>
  mutate(hours_at_lpv=hours_at_lpv_day3) |>
  mutate(time_lpv_eligible=time_lpv_eligible72) |>
  mutate(day=3)


#Make a Dataset With Up to 3 Observations Per Patient (Day 1,2,3)
df_combined <- day1 |>
  rbind(day2) |> rbind(day3)

#CHECK To SEE IF THERE is a Significant Interaction Between Day and ICU CUlture Score, IF NOT will Favor the Stratified Analyses
model_form <- hours_at_lpv ~ mean_unit_culture_scaled*factor(day) + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible)) + (1 | study_icu) + (1|cohort_id)

culture_mv_mod <- glmmTMB(model_form,
          data = df_combined, 
          family=nbinom2)
summary(culture_mv_mod)
exp(confint(culture_mv_mod))

#Run Models with Patients Nested within ICUs
#Culture ICU
#Create the Multivariable Model Form
model_form <- hours_at_lpv ~ mean_unit_culture_scaled:factor(day) + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible)) + (1 | study_icu) + (1|cohort_id)

culture_mv_mod <- glmmTMB(model_form,
          data = df_combined, 
          family=nbinom2)
summary(culture_mv_mod)
exp(confint(culture_mv_mod))

#ICU Stress
model_form <- hours_at_lpv ~ mean_unit_stress_scaled:factor(day) + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible)) + (1 | study_icu) + (1|cohort_id)

stress_mv_mod <- glmmTMB(model_form,
          data = df_combined, 
          family=nbinom2)
summary(stress_mv_mod)
exp(confint(stress_mv_mod))

#ICU Resources
model_form <- hours_at_lpv ~ mean_unit_resources_scaled:factor(day) + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible)) + (1 | study_icu) + (1|cohort_id)

resource_mv_mod <- glmmTMB(model_form,
          data = df_combined, 
          family=nbinom2)
summary(resource_mv_mod)
exp(confint(resource_mv_mod))


#ICU ORIC
model_form <- hours_at_lpv ~ mean_unit_oric_scaled:factor(day) + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible)) + (1 | study_icu) + (1|cohort_id)

oric_mv_mod <- glmmTMB(model_form,
          data = df_combined, 
          family=nbinom2)
summary(oric_mv_mod)
exp(confint(oric_mv_mod))

#ICU Efficacy
model_form <- hours_at_lpv ~ mean_unit_efficacy_scaled:factor(day) + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible)) + (1 | study_icu) + (1|cohort_id)

efficacy_mv_mod <- glmmTMB(model_form,
          data = df_combined, 
          family=nbinom2)
summary(efficacy_mv_mod)
exp(confint(efficacy_mv_mod))
```


```{r Sensitivity Analysis 1 - LPV Deliverin First 72 Hours}
#Can take the DF-Combined Dataset which has up to 3 rows for each day and sum the total LPV hours and total LPV time 
df <- df_combined |>
  group_by(cohort_id) |>
  mutate(total_lpv_hours=sum(hours_at_lpv, na.rm =TRUE)) |>
  mutate(total_lpv_time_eligible=sum(time_lpv_eligible, na.rm=TRUE)) |>
  filter(row_number()==1) |>
  ungroup()

#Repeat Analyses
#Culture ICU
#Create the Multivariable Model Form
model_form <- total_lpv_hours ~ mean_unit_culture_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(total_lpv_time_eligible)) + (1 | study_icu)

culture_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(culture_mv_mod)
exp(confint(culture_mv_mod))

#ICU Stress
model_form <- total_lpv_hours ~ mean_unit_stress_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(total_lpv_time_eligible)) + (1 | study_icu)

stress_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(stress_mv_mod)
exp(confint(stress_mv_mod))

#ICU Resources
model_form <- total_lpv_hours ~ mean_unit_resources_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(total_lpv_time_eligible)) + (1 | study_icu)

resource_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(resource_mv_mod)
exp(confint(resource_mv_mod))

#ICU ORIC
model_form <- total_lpv_hours ~ mean_unit_oric_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(total_lpv_time_eligible)) + (1 | study_icu)

oric_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(oric_mv_mod)
exp(confint(oric_mv_mod))

#ICU Efficacy
model_form <- total_lpv_hours ~ mean_unit_efficacy_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(total_lpv_time_eligible)) + (1 | study_icu)

efficacy_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(efficacy_mv_mod)
exp(confint(efficacy_mv_mod))
```



```{r Visualize Change in Tidal Volume Over Time}
df_long <- lpv_analytic |>
  group_by(study_icu) |>
  mutate(mean_lpv_day1=mean(percent_lpv, na.rm = TRUE)) |>
  mutate(mean_lpv_day2=mean(percent_lpv_day2, na.rm = TRUE)) |>
  mutate(mean_lpv_day3=mean(percent_lpv_day3, na.rm = TRUE)) |>
  mutate(tv_weighted_day1=mean(tv_weighted24, na.rm = TRUE)) |>
  mutate(tv_weighted_day2=mean(tv_weighted_day2, na.rm = TRUE)) |>
  mutate(tv_weighted_day3=mean(tv_weighted_day3, na.rm = TRUE)) |>
  mutate(n_obs_day1=sum(!is.na(percent_lpv))) |>
  mutate(n_obs_day2=sum(!is.na(percent_lpv_day2))) |>
  mutate(n_obs_day3=sum(!is.na(percent_lpv_day3))) |>
  mutate(tv_mean=mean(tv_weighted24, na.rm = TRUE)) |>
  filter(row_number()==1) |>
  ungroup() |>
  pivot_longer(cols = c(mean_lpv_day1, mean_lpv_day2, mean_lpv_day3, 
                        tv_weighted_day1, tv_weighted_day2, tv_weighted_day3, 
                        n_obs_day1, n_obs_day2, n_obs_day3),
               names_to = "lpv_var", 
               values_to = "value") |>
  select(lpv_var, value, study_icu, tv_mean) |>
  mutate(day = fcase(
    grepl('day1', lpv_var), 1,
    grepl('day2', lpv_var), 2,
    grepl('day3', lpv_var), 3
  )) |>
  mutate(tidal_volume=fcase(
    grepl('tv', lpv_var), round(value, digits = 1)
  )) |>
  mutate(percent_lpv=fcase(
    grepl('lpv', lpv_var), round(value, digits = 2))) |>
  mutate(n_obs=fcase(
    grepl('n_obs', lpv_var), value
  )) |>
  group_by(study_icu, day) |>
  fill(n_obs, .direction = 'updown') |>
  ungroup()

#Look at Tidal Volume Trajectory Between Days 1-3 in Each ICU
ggplot(df_long, aes(x = reorder(study_icu, tv_mean), 
                    y = tidal_volume, color = factor(day), size=n_obs)) +
  geom_point(position = position_dodge(width = 0.65)) +  # Use position_dodge to separate points by day
  scale_x_discrete(expand = expansion(mult = c(0.2, 0.2))) +  # Add space between study_icu categories
  scale_color_viridis_d(option = 'G') +   # Use discrete version for color mapping
  labs(x = "ICU Ranked by Day1 Time-weighted Tidal Volume", y = "Tidal Volume (kg-PBW)", color = "Day") +
  scale_y_continuous(breaks=seq(6.0,8.0, by=0.5), labels=c('6.0', '6.5', '7.0', '7.5', '8.0'), limits = c(6,8)) +
  theme_minimal() +
  scale_x_discrete(labels=c('1','2', '3','4','5','6','7','8',
                            '9','10','11','12','13','14')) +
  guides(size = guide_legend(title = 'Number of Patients'),
        colour = guide_legend(override.aes = list(size=3)))
ggsave('tv_day1-3_byicu.pdf',
       path = 'graphs/')
```

```{r Secondary Analysis - Models with Interaction or Stratified by Mild-Moderate or Severe Baseline Hypoxemia}
#GLOBAL ARDS Definition Severe Cutoff is S/F <= 148
df_severe <- lpv_analytic |>
  mutate(hours_at_lpv=round(hours_at_lpv, digits =0)) |>
  filter(sf_ratio<=148)

df_moderate <- lpv_analytic |>
  mutate(hours_at_lpv=round(hours_at_lpv, digits =0)) |>
  filter(sf_ratio>148)

df <- lpv_analytic |>
  mutate(hours_at_lpv=round(hours_at_lpv, digits =0)) |>
  mutate(severe_hypoxemia=fifelse(sf_ratio<=148, 1, 0))

#IF there is a Positive Interaction Can Use Models with Interaction Term if NOt Can Report Stratified 
model_form <- hours_at_lpv ~ mean_unit_culture_scaled*severe_hypoxemia + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

culture_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(culture_mv_mod)
exp(confint(culture_mv_mod))


#SEVERE MODELS
#Culture ICU
#Create the Multivariable Model Form
model_form <- hours_at_lpv ~ mean_unit_culture_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

culture_mv_mod <- glmmTMB(model_form,
          data = df_severe, 
          family=nbinom2)
summary(culture_mv_mod)
exp(confint(culture_mv_mod))

#ICU Stress
model_form <- hours_at_lpv ~ mean_unit_stress_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

stress_mv_mod <- glmmTMB(model_form,
          data = df_severe, 
          family=nbinom2)
summary(stress_mv_mod)
exp(confint(stress_mv_mod))

#ICU Resources
model_form <- hours_at_lpv ~ mean_unit_resources_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

resource_mv_mod <- glmmTMB(model_form,
          data = df_severe, 
          family=nbinom2)
summary(resource_mv_mod)
exp(confint(resource_mv_mod))

#ICU ORIC
model_form <- hours_at_lpv ~ mean_unit_oric_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

oric_mv_mod <- glmmTMB(model_form,
          data = df_severe, 
          family=nbinom2)
summary(oric_mv_mod)
exp(confint(oric_mv_mod))

#ICU Efficacy
model_form <- hours_at_lpv ~ mean_unit_efficacy_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

efficacy_mv_mod <- glmmTMB(model_form,
          data = df_severe, 
          family=nbinom2)
summary(efficacy_mv_mod)
exp(confint(efficacy_mv_mod))
```

```{r MILD to Moderate Models}
#Culture ICU
#Create the Multivariable Model Form
model_form <- hours_at_lpv ~ mean_unit_culture_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

culture_mv_mod <- glmmTMB(model_form,
          data = df_moderate, 
          family=nbinom2)
summary(culture_mv_mod)
exp(confint(culture_mv_mod))

#ICU Stress
model_form <- hours_at_lpv ~ mean_unit_stress_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

stress_mv_mod <- glmmTMB(model_form,
          data = df_moderate, 
          family=nbinom2)
summary(stress_mv_mod)
exp(confint(stress_mv_mod))

#ICU Resources
model_form <- hours_at_lpv ~ mean_unit_resources_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

resource_mv_mod <- glmmTMB(model_form,
          data = df_moderate, 
          family=nbinom2)
summary(resource_mv_mod)
exp(confint(resource_mv_mod))

#ICU ORIC
model_form <- hours_at_lpv ~ mean_unit_oric_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

oric_mv_mod <- glmmTMB(model_form,
          data = df_moderate, 
          family=nbinom2)
summary(oric_mv_mod)
exp(confint(oric_mv_mod))

#ICU Efficacy
model_form <- hours_at_lpv ~ mean_unit_efficacy_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

efficacy_mv_mod <- glmmTMB(model_form,
          data = df_moderate, 
          family=nbinom2)
summary(efficacy_mv_mod)
exp(confint(efficacy_mv_mod))
gc()
```

#SENSITIVITY ANALYSIS 1: Restrict to 2024
```{r SEnsitivity Analysis- Restrict to 6 Month Period}
lpv_2024 <- lpv_analytic |>
  filter(final_admit_date>='2024-01-01')

#Describe this Smaller Cohort
to_tab <- c('age_at_admission', 
         'gender',
         'female',
         'race', 
         'ethnicity', 
         'nonwhite',
         'admit_year',
         'hospital',
         'icu_type',
         'icu_beds',
         'mv_volume_month',
         'transfer_to_jhhs', 
         'ed_to_icu',
         'ed_to_or',
         'elective_surg',
         'direct_icu',
         'admit_category',
         'elixhauser_count',
         'code_stat_time0',
         'CMR_Index_Mortality',
         'ccsr_system_cat',
         'dx_category',
         'covid_ccsr_dx',
         'covid_icd10_poa',
         'height_inches',
         'weight_kg',
         'bmi',
         'max_nr_sofa',
         'pf_ratio',
         'sf_ratio',
         'pf_fio2',
         'sf_fio2',
         'resp_compliance_baseline',
         'first_vent_mode',
         'received_ecmo',
         'inhosp_death',
         'dc_hospice')

factors_tab <- c(
         'gender',
         'female',
         'race', 
         'ethnicity', 
         'nonwhite',
         'admit_year',
         'hospital',
         'icu_type',
         'transfer_to_jhhs', 
         'admit_category',
         'ed_to_icu',
         'ed_to_or',
         'elective_surg',
         'direct_icu',
         'ccsr_system_cat',
         'dx_category',
         'covid_ccsr_dx',
         'covid_icd10_poa',
         'first_vent_mode',
         'received_ecmo',
         'code_stat_time0',
         'inhosp_death',
         'dc_hospice')

nonnorm_tab <-c('age_at_admission', 
          'icu_beds',
         'mv_volume_month',
         'elixhauser_count',
         'CMR_Index_Mortality',
         'height_inches',
         'weight_kg',
         'bmi',
         'max_nr_sofa',
         'pf_ratio',
         'sf_ratio',
         'pf_fio2',
         'sf_fio2',
         'resp_compliance_baseline')

#First Create Table with Numbers of Missing and Min/Max
tab1 <- CreateTableOne(vars = to_tab, data=lpv_2024, factorVars = factors_tab)
summary(tab1)
#Now Create a Traditional Table 1, Using Non-parametric Testing
tab1 <- CreateTableOne(vars = to_tab, strata="study_icu", data=lpv_2024, factorVars = factors_tab, addOverall = TRUE)
print(tab1, nonnormal=nonnorm_tab)
tab1_excel <- print(tab1, nonnorm=nonnorm_tab, printToggle = FALSE)
write.csv(tab1_excel, file="tables/table1_2024ONLY_by-icu.csv")
rm(tab1_excel)
```


```{r PRIMARY ANALYSIS in 2024 Cohort}
#Round Hours
df <- lpv_2024 |>
  mutate(hours_at_lpv=round(hours_at_lpv, digits = 0))

#Culture ICU
#Create the Multivariable Model Form
model_form <- hours_at_lpv ~ mean_unit_culture_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

culture_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(culture_mv_mod)
exp(confint(culture_mv_mod))

#ICU Stress
model_form <- hours_at_lpv ~ mean_unit_stress_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

stress_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(stress_mv_mod)
exp(confint(stress_mv_mod))

#ICU Resources
model_form <- hours_at_lpv ~ mean_unit_resources_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

resource_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(resource_mv_mod)
exp(confint(resource_mv_mod))

#ICU ORIC
model_form <- hours_at_lpv ~ mean_unit_oric_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

oric_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(oric_mv_mod)
exp(confint(oric_mv_mod))

#ICU Efficacy
model_form <- hours_at_lpv ~ mean_unit_efficacy_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

efficacy_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(efficacy_mv_mod)
exp(confint(efficacy_mv_mod))
```

```{r Primary Analysis - Exclude Mixed Med/Surg ICUs}
#Round Hours
df <- lpv_analytic |>
  mutate(hours_at_lpv=round(hours_at_lpv, digits = 0)) |>
  filter(icu_type!='Mixed')

#Culture ICU
#Create the Multivariable Model Form
model_form <- hours_at_lpv ~ mean_unit_culture_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

culture_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(culture_mv_mod)
exp(confint(culture_mv_mod))

#ICU Stress
model_form <- hours_at_lpv ~ mean_unit_stress_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

stress_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(stress_mv_mod)
exp(confint(stress_mv_mod))

#ICU Resources
model_form <- hours_at_lpv ~ mean_unit_resources_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

resource_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(resource_mv_mod)
exp(confint(resource_mv_mod))

#ICU ORIC
model_form <- hours_at_lpv ~ mean_unit_oric_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

oric_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(oric_mv_mod)
exp(confint(oric_mv_mod))

#ICU Efficacy
model_form <- hours_at_lpv ~ mean_unit_efficacy_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

efficacy_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(efficacy_mv_mod)
exp(confint(efficacy_mv_mod))

```

# Use Marginal Effects Package to Generate Predictions for Graphs with Adjusted Results
```{r Re-Run Multivariable Models Used in Primary Analysis}
#Round Hours
df <- lpv_analytic |>
  mutate(hours_at_lpv=round(hours_at_lpv, digits = 0))

#Culture ICU
#Create the Multivariable Model Form
model_form <- hours_at_lpv ~ mean_unit_culture_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

culture_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(culture_mv_mod)
exp(confint(culture_mv_mod))

#ICU Stress
model_form <- hours_at_lpv ~ mean_unit_stress_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

stress_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(stress_mv_mod)
exp(confint(stress_mv_mod))

#ICU Resources
model_form <- hours_at_lpv ~ mean_unit_resources_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

resource_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(resource_mv_mod)
exp(confint(resource_mv_mod))

#ICU ORIC
model_form <- hours_at_lpv ~ mean_unit_oric_scaled + age_at_admission + female + nonwhite + CMR_Index_Mortality + weight_kg + height_inches + covid_icd10_poa + max_nr_sofa + factor(norepi_equivalent) + resp_compliance_baseline + sf_ratio + admit_category + study_month + mv_volume_month + icu_type + icu_beds + factor(hospital) + ccsr_system_cat + offset(log(time_lpv_eligible24)) + (1 | study_icu)

oric_mv_mod <- glmmTMB(model_form,
          data = df, 
          family=nbinom2)
summary(oric_mv_mod)
exp(confint(oric_mv_mod))

```

```{r Function to Generate Predictions Over the Range of Unit Climate/Culture Scores}
#Generates 20 Equally Spaced Values from the Minimum to Maximum of Each Survey Score
function(score) {
  if (all(is.na(score))) {
    stop("The 'score' variable contains only missing values.")
  }
  
  # Ensure score is numeric and handle NA values
  scale <- seq(
    from = range(score, na.rm = TRUE)[1],  # Start of range
    to = range(score, na.rm = TRUE)[2],    # End of range
    length.out = 20                        # Number of points
  )
  
  return(scale)
}

#Culture Predictions
culture_scale <- scale_function(lpv_analytic$mean_unit_culture_scaled)

p <- predictions(
  model= culture_mv_mod, 
  newdata = datagrid(
        mean_unit_culture_scaled = culture_scale,
        female = 0,
        nonwhite = 0,
        CMR_Index_Mortality = median(df$CMR_Index_Mortality, na.rm = TRUE),
        weight_kg = median(df$weight_kg, na.rm = TRUE),
        height_inches = median(df$height_inches, na.rm = TRUE),
        covid_icd10_poa = 0,
        max_nr_sofa = median(df$max_nr_sofa, na.rm = TRUE),
        norepi_equivalent = 0,
        resp_compliance_baseline = median(df$resp_compliance_baseline, na.rm = TRUE),
        sf_ratio = median(df$sf_ratio, na.rm = TRUE),
        admit_category = 'ED',
        study_month = median(df$study_month, na.rm = TRUE),
        mv_volume_month = median(df$mv_volume_month, na.rm = TRUE),
        icu_type = 'Medical',
        icu_beds = median(df$icu_beds, na.rm = TRUE),
        hospital = 'JHH',
        ccsr_system_cat = 'Circulatory'),
      re.form = NA
  )

culture_predict <- data.frame(
  mean_unit_culture = p$mean_unit_culture_scaled * 
    sd(lpv_analytic$mean_unit_culture, na.rm = TRUE),
  percent_lpv = p$estimate / p$time_lpv_eligible24)


#ICU Stress
stress_scale <- scale_function(lpv_analytic$mean_unit_stress_scaled)

p <- predictions(
  model= stress_mv_mod, 
  newdata = datagrid(
        mean_unit_stress_scaled = stress_scale,
        female = 0,
        nonwhite = 0,
        CMR_Index_Mortality = median(df$CMR_Index_Mortality, na.rm = TRUE),
        weight_kg = median(df$weight_kg, na.rm = TRUE),
        height_inches = median(df$height_inches, na.rm = TRUE),
        covid_icd10_poa = 0,
        max_nr_sofa = median(df$max_nr_sofa, na.rm = TRUE),
        norepi_equivalent = 0,
        resp_compliance_baseline = median(df$resp_compliance_baseline, na.rm = TRUE),
        sf_ratio = median(df$sf_ratio, na.rm = TRUE),
        admit_category = 'ED',
        study_month = median(df$study_month, na.rm = TRUE),
        mv_volume_month = median(df$mv_volume_month, na.rm = TRUE),
        icu_type = 'Medical',
        icu_beds = median(df$icu_beds, na.rm = TRUE),
        hospital = 'JHH',
        ccsr_system_cat = 'Circulatory'),
      re.form = NA
  )

stress_predict <- data.frame(
  mean_unit_stress = p$mean_unit_stress_scaled * 
    sd(lpv_analytic$mean_unit_stress, na.rm = TRUE),
  percent_lpv = p$estimate / p$time_lpv_eligible24)

#ICU Resources
resources_scale <- scale_function(lpv_analytic$mean_unit_resources_scaled)

p <- predictions(
  model= resource_mv_mod, 
  newdata = datagrid(
        mean_unit_resources_scaled = resources_scale,
        female = 0,
        nonwhite = 0,
        CMR_Index_Mortality = median(df$CMR_Index_Mortality, na.rm = TRUE),
        weight_kg = median(df$weight_kg, na.rm = TRUE),
        height_inches = median(df$height_inches, na.rm = TRUE),
        covid_icd10_poa = 0,
        max_nr_sofa = median(df$max_nr_sofa, na.rm = TRUE),
        norepi_equivalent = 0,
        resp_compliance_baseline = median(df$resp_compliance_baseline, na.rm = TRUE),
        sf_ratio = median(df$sf_ratio, na.rm = TRUE),
        admit_category = 'ED',
        study_month = median(df$study_month, na.rm = TRUE),
        mv_volume_month = median(df$mv_volume_month, na.rm = TRUE),
        icu_type = 'Medical',
        icu_beds = median(df$icu_beds, na.rm = TRUE),
        hospital = 'JHH',
        ccsr_system_cat = 'Circulatory'),
      re.form = NA
  )

resources_predict <- data.frame(
  mean_unit_resources = p$mean_unit_resources_scaled * 
    sd(lpv_analytic$mean_unit_resources, na.rm = TRUE),
  percent_lpv = p$estimate / p$time_lpv_eligible24)

#ICU ORIC Score
oric_scale <- scale_function(lpv_analytic$mean_unit_oric_scaled)

p <- predictions(
  model= oric_mv_mod, 
  newdata = datagrid(
        mean_unit_oric_scaled = oric_scale,
        female = 0,
        nonwhite = 0,
        CMR_Index_Mortality = median(df$CMR_Index_Mortality, na.rm = TRUE),
        weight_kg = median(df$weight_kg, na.rm = TRUE),
        height_inches = median(df$height_inches, na.rm = TRUE),
        covid_icd10_poa = 0,
        max_nr_sofa = median(df$max_nr_sofa, na.rm = TRUE),
        norepi_equivalent = 0,
        resp_compliance_baseline = median(df$resp_compliance_baseline, na.rm = TRUE),
        sf_ratio = median(df$sf_ratio, na.rm = TRUE),
        admit_category = 'ED',
        study_month = median(df$study_month, na.rm = TRUE),
        mv_volume_month = median(df$mv_volume_month, na.rm = TRUE),
        icu_type = 'Medical',
        icu_beds = median(df$icu_beds, na.rm = TRUE),
        hospital = 'JHH',
        ccsr_system_cat = 'Circulatory'),
      re.form = NA
  )

oric_predict <- data.frame(
  mean_unit_oric = p$mean_unit_oric_scaled * 
    sd(lpv_analytic$mean_unit_oric, na.rm = TRUE),
  percent_lpv = p$estimate / p$time_lpv_eligible24)

```


```{r Graph the Predicted Marginal Effects on the Scatter Plots}
df <- lpv_analytic |> 
  group_by(study_icu) |>
  mutate(mean_lpv=mean(percent_lpv, na.rm = TRUE)) |>
  mutate(mean_lpv_restrict=mean(percent_lpv_restrict, na.rm = TRUE)) |>
  mutate(icu_count=n()) |>
  ungroup() |>
  group_by(icu_type) |> 
  mutate(count_type=sum(icu_count)) |> 
  ungroup()

#Scatter Plots With Predicted Line Added - Generated from Predicted Hours of LPV for the Median Patient at JHH in Medical ICU
#Culture
ggplot() +
  geom_smooth(data = df, 
              aes(x=mean_unit_culture, 
                  y=percent_lpv), 
              method = 'glm', alpha = 0.5) +
  geom_smooth(data=culture_predict, 
              aes(x=mean_unit_culture, y=percent_lpv), 
              method = 'glm', alpha = 0.5, se = FALSE,
              linetype = 'dashed', color = 'red') +
  geom_point(data=df,
             aes(x=mean_unit_culture, y=mean_lpv, 
                 size = icu_count, 
                 color = reorder(icu_type, -count_type)), 
             alpha = 0.8) +
  scale_y_continuous(breaks=seq(0,1, by=0.1), labels = scales::percent, limits = c(0,1.0)) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ICU Implementation Culture Score') +
  ylab('% of Time LPV Delivered, First 24 Hours') +
  guides(colour = guide_legend(override.aes = list(size=5), title = 'ICU Type'), size = 'none')
ggsave('scatter_marginaleffect_culture.pdf',
       path = 'graphs/')

#Stress
ggplot() +
  geom_smooth(data = df, 
              aes(x=mean_unit_stress, 
                  y=percent_lpv), 
              method = 'glm', alpha = 0.5) +
  geom_smooth(data=stress_predict, 
              aes(x=mean_unit_stress, y=percent_lpv), 
              method = 'glm', alpha = 0.5, se = FALSE,
              linetype = 'dashed', color = 'red') +
  geom_point(data=df,
             aes(x=mean_unit_stress, y=mean_lpv, 
                 size = icu_count, 
                 color = reorder(icu_type, -count_type)), 
             alpha = 0.8) +
  scale_y_continuous(breaks=seq(0,1, by=0.1), labels = scales::percent, limits = c(0,1.0)) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ICU Staff Stress Score') +
  ylab('% of Time LPV Delivered, First 24 Hours') +
  guides(colour = guide_legend(override.aes = list(size=5), title = 'ICU Type'), size = 'none')
ggsave('scatter_marginaleffect_stress.pdf',
       path = 'graphs/')

#Resources
ggplot() +
  geom_smooth(data = df, 
              aes(x=mean_unit_resources, 
                  y=percent_lpv), 
              method = 'glm', alpha = 0.5) +
  geom_smooth(data=resources_predict, 
              aes(x=mean_unit_resources, y=percent_lpv), 
              method = 'glm', alpha = 0.5, se = FALSE,
              linetype = 'dashed', color = 'red') +
  geom_point(data=df,
             aes(x=mean_unit_resources, y=mean_lpv, 
                 size = icu_count, 
                 color = reorder(icu_type, -count_type)), 
             alpha = 0.8) +
  scale_y_continuous(breaks=seq(0,1, by=0.1), labels = scales::percent, limits = c(0,1.0)) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ICU Resources Score') +
  ylab('% of Time LPV Delivered, First 24 Hours') +
  guides(colour = guide_legend(override.aes = list(size=5), title = 'ICU Type'), size = 'none')
ggsave('scatter_marginaleffect_resources.pdf',
       path = 'graphs/')

#ORIC
ggplot() +
  geom_smooth(data = df, 
              aes(x=mean_unit_oric, 
                  y=percent_lpv), 
              method = 'glm', alpha = 0.5) +
  geom_smooth(data=oric_predict, 
              aes(x=mean_unit_oric, y=percent_lpv), 
              method = 'glm', alpha = 0.5, se = FALSE,
              linetype = 'dashed', color = 'red') +
  geom_point(data=df,
             aes(x=mean_unit_oric, y=mean_lpv, 
                 size = icu_count, 
                 color = reorder(icu_type, -count_type)), 
             alpha = 0.8) +
  scale_y_continuous(breaks=seq(0,1, by=0.1), labels = scales::percent, limits = c(0,1.0)) +
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  xlab('ICU ORIC Score') +
  ylab('% of Time LPV Delivered, First 24 Hours') +
  guides(colour = guide_legend(override.aes = list(size=5), title = 'ICU Type'), size = 'none')
ggsave('scatter_marginaleffect_oric.pdf',
       path = 'graphs/')
```

```{r Get Legend}
ggplot(df) +
  # First geom_smooth with alpha aesthetic in the legend
  geom_smooth(aes(x = mean_unit_culture, 
                  y = percent_lpv, 
                  alpha = "Unadjusted Regression Line"),  # Label for alpha legend
              method = 'glm') +
  
  # Second geom_smooth with manual legend addition
  geom_smooth(data = culture_predict, 
              aes(x = mean_unit_culture, 
                  y = percent_lpv, 
                  linetype = "Adjusted Model Predictions"),  # Add to legend
              method = 'glm', 
              se = FALSE, 
              color = 'red') +
  
  # Points layer
  geom_point(aes(x = mean_unit_culture, 
                 y = mean_lpv, 
                 size = icu_count, 
                 color = reorder(study_icu, -mean_unit_culture)), 
             alpha = 0.8) +
  
  # Minimal theme and color scale
  theme_minimal() +
  scale_color_viridis_d(option = 'G') +
  
  # Labels
  xlab('ICU Implementation Culture Score') +
  ylab('% of Time LPV Delivered, First 24 Hours') +
  
  # Guides for legends
  guides(
    colour = 'none',  # Hide color legend
    alpha = guide_legend(title = ""),  # Rename alpha legend
    size = guide_legend(title = 'Number of Patients'),
    linetype = guide_legend(title = "")  # Add line type legend
  ) +
  
  # Manual scales for aesthetics
  scale_alpha_manual(values = c("Unadjusted Regression Line" = 0.5)) +
  scale_linetype_manual(values = c("Adjusted Model Predictions" = "dashed"))
ggsave('scatter_marginaleffect_legend.pdf',
       path = 'graphs/')
```


```{r Save LPV Analytic File}

```
